+++ csrc/pythonInterface.cpp
@@ -443,7 +443,7 @@ void cfp8_e5m2_approx_matmul(uint8_t* A, uint8_t* B, float* C, int M, int N, int
     fp8_e5m2_approx_matmul(A, B, C, M, N, K, stream);
 }
 
-void cbf16_approx_matmul(__nv_bfloat16* A, __nv_bfloat16* B, float* C, int M, int N, int K, cudaStream_t stream) {
+void cbf16_approx_matmul(__nv_bfloat16* A, __nv_bfloat16* B, __nv_bfloat16* C, int M, int N, int K, cudaStream_t stream) {
     bf16_approx_matmul(A, B, C, M, N, K, stream);
 }
